+++ src/languages/typescript.js
@@ -25,7 +25,8 @@ export default function(hljs) {
       'TypeError URIError Number Math Date String RegExp Array Float32Array ' +
       'Float64Array Int16Array Int32Array Int8Array Uint16Array Uint32Array ' +
       'Uint8Array Uint8ClampedArray ArrayBuffer DataView JSON Intl arguments require ' +
-      'module console window document any number boolean string void Promise'
+      'module console window document any number boolean string void Promise ' +
+      'setInterval setTimeout clearInterval clearTimeout'
   };
   var DECORATOR = {
     className: 'meta',